[util] move LatencyPipe into util
