+++ src/main/scala/rocketchip/TestHarness.scala
@@ -5,6 +5,7 @@ package rocketchip
 import Chisel._
 import cde.{Parameters, Field}
 import rocket.Util._
+import util.LatencyPipe
 import junctions._
 import junctions.NastiConstants._
 
@@ -173,25 +174,5 @@ class JTAGVPI(implicit val p: Parameters) extends BlackBox {
     // Success is determined by the gdbserver
     // which is controlling this simulation.
     tbsuccess := Bool(false)
-  }
-}
-
-class LatencyPipe[T <: Data](typ: T, latency: Int) extends Module {
-  val io = new Bundle {
-    val in = Decoupled(typ).flip
-    val out = Decoupled(typ)
-  }
-
-  def doN[T](n: Int, func: T => T, in: T): T =
-    (0 until n).foldLeft(in)((last, _) => func(last))
-
-  io.out <> doN(latency, (last: DecoupledIO[T]) => Queue(last, 1, pipe=true), io.in)
-}
-
-object LatencyPipe {
-  def apply[T <: Data](in: DecoupledIO[T], latency: Int): DecoupledIO[T] = {
-    val pipe = Module(new LatencyPipe(in.bits, latency))
-    pipe.io.in <> in
-    pipe.io.out
   }
 }
